Move complete module activation logic to a hook.
